CryptoPkg: Fix linker warnings related to incompatible types
LTO builds do stricter type checking than no-LTO builds and it complains
about several files that use incompatible declarations.

struct tm was fine, it just had to be removed after including
<CrtLibSupport.h> because that file contains the same declaration.

+++ CryptoPkg/Library/MbedTlsLib/CrtWrapper.c
@@ -18,7 +18,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 int
 my_snprintf (
   char        *str,
-  size_t      size,
+  long long   size,
   const char  *format,
   ...
   )
